requirements.txt
django==4.2
python-decouple
psycopg[binary]
djangorestframework
markdown
django-filter
drf-yasg
Pillow
djangorestframework-simplejwt
faker
requests
django-filter
gunicorn
black
flake8